feat(todo): add create todo functionality
POST :host/api/todo with the corresponding json now is able to write to
the database and return a full Todo json. included in this commit to
make this functional and "type safe" include
- change to `parseRequestBody` to take in a type parameter: this allows
the type checker to be able to validate what is coming out of it.
- s/bodyParser.ts/parseRequestBody.ts/
- explicitly set todo.description to null if not present
- s/`getDb`/`getDB`/g for style reasons
- implement `TodoRepoPSQL.todo_create`
- add `todo_tag` link table to `Database` type
- change `Todo` model
  - `Todo.completed` now optional to reflect it will be filled in later
  - `InsertTodo.tags` now uses `InsertTodo | string` to reflect the tags
  may not exist
  - `Todo.tags` moved to top to make the type checker happy (i think)
- inline return val in `TodoService.create_todo` for style
- migrate database
  - add unique constraint to `tag.name` under name "tag_unique"
  - add primary key constraint to `todo_tag.todo_id` and
  "todo_tag.tag_id" under name "todo_tag_pkey"

should this have been a set of commits? probably