fix(todo): check id exists on repo
read and update used to take the ID as valid in all cases,
now a check is performed using a new predicate `todo_does_exist` these
methods import the todo repo factory, check if the ID exists and then
throws a new error which propagates up to the response handler.
This allows for better error messages and eventually proper Errors and
status codes.